% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{taskscheduleR-package}
\alias{taskscheduleR-package}
\alias{taskscheduleR}
\title{Schedule R Scripts/Processes with the Windows Task Scheduler}
\description{
Schedule R scripts/processes with the Windows task scheduler. This allows R users to automate R processes on specific timepoints from R itself.
The package is basically a wrapper around the Schtasks.exe functionality. 
More information about schtasks can be found at https://msdn.microsoft.com/en-us/library/windows/desktop/bb736357%28v=vs.85%29.aspx
or at the doc folder inside this package.
}
