% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskscheduleR.R
\name{taskscheduler_ls}
\alias{taskscheduler_ls}
\title{Get all the tasks which are currently scheduled at the Windows task scheduler.}
\usage{
taskscheduler_ls(author)
}
\arguments{
\item{author}{a character vector of authors (the Author column) indicating who created the task (e.g. admin, yourusername).
Defaults to all authors.}
}
\value{
a data.frame with scheduled tasks as returned by schtasks /Query
}
\description{
Get all the tasks which are currently scheduled at the Windows task scheduler.
}
\examples{
x <- taskscheduler_ls()
x
}

