% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskschedulerAddin.R
\name{taskschedulerAddin}
\alias{taskschedulerAddin}
\title{Launch an RStudio addin which allows to schedule an Rscript interactively.}
\usage{
taskschedulerAddin(RscriptRepository, debug = TRUE)
}
\arguments{
\item{RscriptRepository}{path to the folder where R scripts will be copied to and launched. Defaults to the extdata folder in the taskscheduleR R library}

\item{debug}{passed on to \code{\link{taskscheduler_create}}}
}
\value{
the return of \code{\link[shiny]{runGadget}}
}
\description{
Launch an RStudio addin which allows to schedule an Rscript interactively.
}
\examples{
\dontrun{
taskschedulerAddin()
}
}
